requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
invisible-watermark>=0.2.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
datasets>=2.14.0
huggingface-hub>=0.17.0
timm>=0.9.0
scikit-image>=0.21.0
tqdm>=4.66.0